#77 - Split test class.
