+++ src/main/java/com/vzurauskas/nereides/jackson/SmartJson.java
@@ -3,7 +3,6 @@ package com.vzurauskas.nereides.jackson;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-
 import java.io.InputStream;
 import java.util.Arrays;
 import java.util.Optional;
